Fix --no-comment behavior.
The --no-comment parameter behaved incorrectly, doing the opposite
of what it was supposed to do;  that is, specifying it caused the
comment to appear, and not using it did not.  This has been fixed.

In addition, the short form of this, -Q, was not recognized by
getopt_long().  This has also been fixed.