Add support for custom comments.
You had the option to remove the user agent from the comment section
of the packet.  Now, you can finally specify your own comment using
the `-M` option.

APRS packet comments are limited to 28-43 characters, depending on
the packet type.  This change allows for up to 43 characters, with a
warning that anything longer might be truncated.

Note that if the `-Q`/`--no-comment` option is specified, this will
override it.