Create C89 unofficial patch.
I don't know why I think I can port this to MS-DOS or Windows 3.1,
but just in case someone out there wants to get this app to build
on a C compiler that only supports the C89 standard, here you go.

This is untested, extremely unsupported, full of buffer overflows,
and requires you to give up APRS-IS support.  Maintenance will be
provided on a "whenever I feel like it" schedule.