Add preliminary support for running under DOS
Why?  I was bored.  This commit adds support for compiling this app
under FreeDOS, MS-DOS, or a DOS-compatible system.  You will need
to use the OpenWatcom toolchain to compile and specify the `-D_DOS`
compile-time flag.  APRS-IS will be automatically disabled.

Note that this will not work.  I am still looking for a version of
getopt.h that is compatible with DOS and has a GPL-compatible
license.  The files getoptfd.h and getoptfd.c have intentionally
not been uploaded.