Add support for changing the APRS icon.
This commit allows users to select their own icon, as opposed to the
standard weather icon beind hardcoded.  Note that we do not check to
make sure that the user's input maps to a standards-approved icon, as
I'm sure there are plenty of non-compliant, custom, or "this isn't what
it looks like in my pre-Unicode codepage" icons out there.