Add BUFSIZE check to snprintf_verify().
After hearing Steve "Tiberius" Gibson (of Security NOW fame) explain the
"Citrixbleed" bug, I thought to check better the output of snprintf().
Well, I already have a wrapper for it, so I also made sure it did not
exceed the preprocessor macro BUFSIZE.  While that may not indicate a
security issue, it would indicate that we cannot guarantee that output
will be correct and should not continue execution.