Modify build scripts to accomodate Win32.
This commit adds preliminary support for allowing this code to compile
on Microsoft Windows (Vista, Server 2008, and newer).  A new configure
flag, --enable-windows, allows this app to build and link against the
WinSock2 API instead of the standard POSIX stuff.

So far, this has only been tested with MinGW, but I was able to build
Win32 binaries of the previous version, with some coaxing.