Add garbage to files for ISO-compliant compilers.
This commit adds a single statement to aprs-is.c and c99math.c that
will allow compilation on ISO-compliant compilers, such as GCC with
the `-pedantic` or `-pedantic-errors` flags. Under a couple certain
conditions -- specifically, when compiling for non-DOS platforms or
with APRS-IS disabled -- two source files would wind up blank after
preprocessing, which is a violation of the ISO C specifications.

This commit adds a single typedef statement to the aforementioned C
source files that does nothing to affect program flow but will make
even the strictest compilers happy.

In the spirit of writing good and portable code (as much as the GNU
extensions will allow), I've also added `-pedantic-errors` into the
configure.ac script, and by extension, into our Makefile. This will
force future maintainers to write high-quality code.