Fix bug preventing compilation without APRS-IS.
The previous commit about `gcc -pedantic` pointed out this flaw in the
configure.ac file. I'd always supported the `--disable-aprs-is` option
to allow compiling without APRS-IS support (for example, if you wanted
to use this as a library or a helper app).  However, that test did not
succeed.  Now it does.