Make I/O non-blocking, add --timeout parameter.
Thank you to DL9SEC for reporting a strange case.  The IGate server
cwop.aprs2.net has multiple IP addresses defined in DNS, and not all
of them work.  This causes aprs-weather-submit to hang.  This commit
converts all socket operations to non-blocking.

I've also added a new --timeout parameter that can specify how many
seconds the app will wait before trying the next IPv4/IPv6 address,
or failing.  The default is 15 seconds, because that sounded nice.

This does add a few kilobytes to the compiled app, so I may wrap this
in a `#ifdef ONLY_BLOCKING_IO` or something.

I'm still wondering if this is major-enough change to warrant a bump
of the version number to 2.0.0.

Fixes: #12