Fix bug in APRS-IS server login.
The `vers` field in the APRS-IS server login is optional, but if you
choose to use it (`aprs-weather-submit` does), then you need to send
the app name and version, separated by a space. I was using a slash.
My bad.  This commit fixes it.

That being said, I've never had a problem with this app before, so
I'm going to guess that this is a mostly cosmetic change.