Add support for APRS 1.2.1 device type identifier
This commit adds support for the `Z` field, a two-character device type
identifier, as defined in the APRS 1.2.1 `weather-new.txt` addendum by
Bob Bruninga, WB4APR (SK) on March 24, 2011.  Device types are exactly
two characters long.  What they mean is implementation-specific and not
defined in the APRS specification or its addenda, the best I could tell.

Basically, if you don't know what this is, don't use it.  If you do know
what it is, go nuts.