Fix a bug where `Z` is always included.
A logic error in the code sent the device type identifier field
(`Z`) with every single packet, whether the user asked for it
or not.  It'd be prepended to the comment with no value, adding
the letter Z to everything.  This has been corrected.

Format: markdown
Reviewer: github
Tracker: github
Signed-off-by: Colin Cogle <colin@colincogle.name>