Fall back to gmtime() for DOS builds.
In version 2.0, we switched to using `gmtime_r()` from `gmtime()` to
avoid a possible security issue.  However, this resulted in OpenWatcom
failing to build this app.  This is because OpenWatcom only supports
the C99 standard, while `gmtime_r()` is a GNU/POSIX extension.

This commit reverts this behavior on DOS platforms by using conditional
logic to fall back to `gmtime()`.  It does introduce a small security
risk, but without APRS-IS support in DOS at this time, I feel that it
is an acceptable tradeoff.

Fixes-Bug-Introduced-By: 5e1d35af39a6ec91de944112b8019286bd196db0
Format: markdown
Issue: #36
Tracker: github