Warn that custom icons for weather reports are unsupported.
[damso](https://github.com/damso) came to me with a very strange issue.
He set a custom icon in his APRS packets, and suddenly, his weather data
was being interpreted as a comment.  I dug into this, and unfortunately,
I found out that this is perfectly normal behavior per the APRS 1.01
specification.

Thus, custom icons are not supported when your command-line includes
any weather data (except for wind direction and wind speed).  As this
is not a bug with the app itself, I am updating the documentation to
note that use of the `--icon` parameter should be discouraged.  I will
leave it in the app, as I know of at least one person who uses this
`aprs-weather-submit` app for position data, where custom icons are
perfectly fine.

Stupid APRS.

Format: markdown
Milestone: v2.0.1
Issue: #30
Tracker: github